ultrasonic HC-SR04 cal_distance of

+++ STM32Cube_FW_L0_V1.10.0/Projects/STM32L073RZ-Nucleo/stm32_ultrasonic/Src/HC_SR04/hc_sr04.h
@@ -41,7 +41,7 @@ int timing(void);
 	floart centigrade,当前环境温度，单位：摄氏度
 	return:距离，单位：mm
 */
-double cal_distance(int us, double centigrade);
+float cal_distance(double us, double centigrade);
 
 int timer_init(void);
 /*
